Split minion
